refactor/#86: parent/mission 패키지 수정
