+++ app/src/main/java/com/kiero/data/parent/mission/repositoryimpl/AutoMissionRepositoryImpl.kt
@@ -1,13 +1,13 @@
-package com.kiero.data.mission.repositoryimpl
+package com.kiero.data.parent.mission.repositoryimpl
 
 import com.kiero.core.common.util.suspendRunCatching
-import com.kiero.data.mission.model.MissionCompleteModel
-import com.kiero.data.mission.model.MissionSuggestionModel
-import com.kiero.data.mission.model.SuggestedMissionModel
-import com.kiero.data.mission.model.toModel
-import com.kiero.data.mission.remote.datasource.AutoMissionDataSource
-import com.kiero.data.mission.remote.dto.request.MissionCreateDto
-import com.kiero.data.mission.repository.AutoMissionRepository
+import com.kiero.data.parent.mission.remote.datasource.AutoMissionDataSource
+import com.kiero.data.parent.mission.remote.dto.request.MissionCreateDto
+import com.kiero.domain.entity.parent.mission.MissionCompleteModel
+import com.kiero.domain.entity.parent.mission.MissionSuggestionModel
+import com.kiero.domain.entity.parent.mission.SuggestedMissionModel
+import com.kiero.domain.entity.parent.mission.toModel
+import com.kiero.domain.repository.parent.mission.AutoMissionRepository
 import kotlinx.coroutines.withTimeout
 import javax.inject.Inject
