+++ app/src/main/java/com/kiero/presentation/parent/schedule/mission/auto/viewmodel/AutoMissionViewModel.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.snapshotFlow
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kiero.core.localstorage.info.UserInfoManager
-import com.kiero.data.mission.model.SuggestedMissionModel
-import com.kiero.data.mission.repository.AutoMissionRepository
+import com.kiero.domain.entity.parent.mission.SuggestedMissionModel
+import com.kiero.domain.repository.parent.mission.AutoMissionRepository
 import com.kiero.presentation.parent.schedule.mission.auto.model.MissionUiModel
 import com.kiero.presentation.parent.schedule.mission.auto.state.AutoMissionSideEffect
 import com.kiero.presentation.parent.schedule.mission.auto.state.AutoMissionState
